consolidate mistake filtering into 'checkStep' wrapper function
