+++ packages/step-checker/src/index.ts
@@ -1,4 +1,5 @@
 import StepChecker from "./step-checker";
+export {checkStep} from "./step-checker";
 export type {Context, Mistake} from "./types";
 
 export default StepChecker;
